Fix some implicit-bool-conversion warnings
PiperOrigin-RevId: 524108216

+++ src/xnnpack/aarch32-assembler.h
@@ -373,8 +373,8 @@ class MemOperand {
   AddressingMode mode() const { return mode_; }
 
   // These are bits used for encoding, named based on the encoding description.
-  int32_t u() { return offset_ >= 0; }
-  int32_t p() { return mode_ != AddressingMode::kPostIndexed; }
+  int32_t u() { return static_cast<int32_t>(offset_ >= 0); }
+  int32_t p() { return static_cast<int32_t>(mode_ != AddressingMode::kPostIndexed); }
   // Note, kPostIndexed will write back, but doesn't need to set bit w.
   int32_t w() { return 0; }
